Queue applications
The 2 applications using queue seen in class: “01” generated all the
words of a given alphabet in increasing order, “labyrinth” found the
shortest exit path in a maze.

We do not have our own implementation of a Queue yet, so this uses
java.util.Queue. Note that the methods “enqueue” and “dequeue” have
different names in that interface: “add” and “remove”.

As a concrete implementation of the Queue, I picked
java.util.LinkedList here.